refactor sonarEngine
