+++ cowfarts/src/engines/sonarEngine/sonarSource.ts
@@ -1,0 +1,3 @@
+export default interface SonarSource {
+  sample: (bearing: number, sampleTime?: number) => number;
+}
