+++ cowfarts/src/engines/sonarEngine/waterfalls/createWaterfall.ts
@@ -1,11 +1,5 @@
-// Idea: We want to be constantly updating this w/ the game state,
-// since these need to be canvases, and constantly updated.
-// In a certain sense, historical sonar data is captured solely in this interface.
-// SO SINGLETON THEN
-
-import { ElemSingleton, show, hide } from "../../../util/elemSingleton";
-import BroadbandSource from "../broadbandSource";
 import { Pipe } from "../../../util/pipe";
+import SonarSource from "../sonarSource";
 
 interface DisplaySettings {
   gain: number;
@@ -38,8 +32,8 @@ const sampleIdxToTime = (sample: number) => {
   return Math.floor(sample * UPDATE_INTERVAL_MS + gameEpoch);
 };
 
-const createBroadbandWaterfall = (
-  source: BroadbandSource,
+const createWaterfall = (
+  source: SonarSource,
   { multiplier, gain, contrast }: DisplaySettings
 ): BroadbandScreen => {
   const pipe = new Pipe<ImageData>();
@@ -94,22 +88,4 @@ const createBroadbandWaterfall = (
   };
 };
 
-const buildWaterfalls = (broadbandSource: BroadbandSource) => ({
-  broadbandShort: createBroadbandWaterfall(broadbandSource, {
-    multiplier: 4,
-    contrast: 1,
-    gain: 0.1,
-  }),
-  broadbandMedium: createBroadbandWaterfall(broadbandSource, {
-    multiplier: 16,
-    contrast: 2,
-    gain: 0.035,
-  }),
-  broadbandLong: createBroadbandWaterfall(broadbandSource, {
-    multiplier: 64,
-    contrast: 4,
-    gain: 0.004,
-  }),
-});
-
-export default buildWaterfalls;
+export default createWaterfall;
